kafka-mongo/api_server/kafka_consumer_container/requirements.txt
kafka-python==2.0.2
pymongo==3.11.4